+++ R/power4test_by_pop_es.R
@@ -325,10 +325,12 @@ c.power4test_by_es <- function(...,
 #' @noRd
 rejection_rates_by_es <- function(object_by_es,
                                   all_columns = FALSE,
+                                  nrep_if_diff = TRUE,
                                   ...) {
   tmpfct <- function(x) {
     out_i <- rejection_rates(x,
                              all_columns = all_columns,
+                             keep_nrep = nrep_if_diff,
                              ...)
     pn <- attr(x, "pop_es_name")
     pv <- attr(x, "pop_es_value")
@@ -351,6 +353,13 @@ rejection_rates_by_es <- function(object_by_es,
                               error = function(e) x[cnames1]))
   out <- do.call(rbind,
                  out)
+  if (nrep_if_diff &&
+      !all_columns) {
+    nrep_range <- try(suppressWarnings(diff(range(out$nrep, na.rm = TRUE))), silent = TRUE)
+    if (isFALSE(nrep_range > 1)) {
+      out$nrep <- NULL
+    }
+  }
   rownames(out) <- NULL
   out
 }